// web/src/types/engine.d.ts
export type EngineWebhookPayload = {
  queueId: string;
  status: string;
  onchainStatus: string;
  queuedAt: string;
  sentAt: string;
  minedAt: string;
  errorMessage: string | null;
  cancelledAt: string | null;
  retryCount: number;
  chainId: number | string;
  fromAddress: string;
  toAddress: string;
  data: string;
  value: string;
  nonce: string | number;
  gasLimit: string | null;
  maxFeePerGas: string | null;
  maxPriorityFeePerGas: string | null;
  gasPrice: string | null;
  transactionType: number;
  transactionHash: string;
  sentAtBlockNumber: number;
  blockNumber: number;
  signerAddress: string | null;
  accountAddress: string | null;
  target: string | null;
  sender: string | null;
  initCode: string | null;
  callData: string | null;
  callGasLimit: string | null;
  verificationGasLimit: string | null;
  preVerificationGas: string | null;
  paymasterAndData: string | null;
  userOpHash: string | null;
  functionName: string | null;
  functionArgs: string | null;
  extension: string | null;
  deployedContractAddress: string | null;
  deployedContractType: string | null;
  onChainTxStatus?: number;
  effectiveGasPrice?: string;
  cumulativeGasUsed?: string;
  accountSalt?: string | null;
  accountFactoryAddress?: string;
  retryGasValues?: string | null;
  retryMaxFeePerGas?: string | null;
  retryMaxPriorityFeePerGas?: string | null;
};